requirements.txt
openai>=1.3.5
streamlit
faiss-cpu
numpy
dotenv